[CASSANDRA-20476] Break up new long running test
